chore: Update dependencies and add Supplier email and phone fields
This commit updates the dependencies in the package.json file of the msgext-northwind-inventory-ts sample. It adds the following dependencies:
- @azure-rest/communication-messages@1.0.1
- @azure/communication-email@1.0.0
- @azure/communication-sms@1.1.0
- @azure/core-auth@1.6.0

Additionally, it modifies the model.ts file to include the SupplierEmail and SupplierPhone fields in the ProductEx interface. This allows for storing and retrieving the email and phone information of suppliers.

These changes were made to improve the functionality and data representation of the Northwind Inventory sample.